[install.sh] Added extra check for discord theme

+++ install.sh
@@ -94,7 +94,8 @@ done
 # Installing discord theme
 for dir in "$HOME/.config"/discord/*.*.*/modules/discord_desktop_core
 do
-    if [ ! -f "$dir/discord-custom.css" ]
+    if [ ! -f "$dir/discord-custom.css" ] && \
+        [ -d "$dir" ]
     then
         ln -sf "$(pwd)/custom/discord/discord-custom.css" "$dir/discord-custom.css"
     fi
